--- delivery-service/src/main/java/com/logistics/platform/delivery_service/delivery/presentation/global/exception/CustomApiException.java
@@ -1,8 +1,0 @@
-package com.logistics.platform.delivery_service.delivery.presentation.global.exception;
-
-public class CustomApiException extends RuntimeException{
-  public CustomApiException(String message) {
-    super(message);
-  }
-
-}
