+++ delivery-service/src/main/java/com/logistics/platform/delivery_service/global/global/exception/CustomApiException.java
@@ -1,4 +1,4 @@
-package com.logistics.platform.delivery_service.deliveryRoute.presentation.global.exception;
+package com.logistics.platform.delivery_service.global.global.exception;
 
 public class CustomApiException extends RuntimeException{
   public CustomApiException(String message) {
